line 14 was modified

+++ math/0x03-probability/poisson.py
@@ -11,7 +11,7 @@ class Poisson:
                 raise TypeError("data must be a list")
             if len(data) <= 2:
                 raise ValueError("data must contain multiple values")
-            self.lambtha = float(sum(data) / len(data))
+            self.lambtha = sum(data) / len(data)
         else:
             if lambtha < 0:
                 raise ValueError("lambtha must be a positive value")
